scripts: restore-etc: retrieve fstab from /flash and copy it at boot time
If there's no fstab on the system and one is present in /flash, we'll use
it. This allows one to build the fstab during the installation process
when hard disks are needed. It's common to ask for partitions during the
first boot, and it is not acceptable to lose them after the reboot. So
being able to keep fstab this way is quite convenient.