scripts: restore-etc: automatically detect ramfs/tmpfs availability
ramfs is always available, but tmpfs is only available if enabled in the
kernel config. On the other hand, tmpfs provides a way to limit the FS
size, while ramfs only does so with old kernel patches.

So we try to use tmpfs by default, and if unavailable, we fall back to ramfs.