scripts: save-etc: maintain a journal of operations on the partition
It is sometimes useful to know whether certain operations succeed or
fail, especially when a system is found to boot on an old config, or
when flash I/O errors are detected.

This patch ensures that any write access to the flash device will be
logged to a file called /.journal on the flash. In case of success,
there's at least one line indicating the success, the file name and
its size. In case of error, maximum information are extracted and
emitted with stderr output and status code of faulty commands, allowing
subtle issues like a failed "mv" or "chmod" to be detected.

The journal format looks like this :

  $date $time $product $model $version $tty $prog: $step: $message

The tty helps detect if the command was issued by hand or via an
interface. "none" is printed if no tty is known. The program reported
will be "save-etc". The step will try to indicate where things fail.
Common ones  will be "archive" and "cancel" (which indicates a rollback
of a change). Note that for all steps except "cancel" and "archive",
only failures are reported, as successes are standard and provide little
information. This ensures that at least one line is emitted on each write
without polluting the output too much.

The message in the case of the "archive" step, the message is composed
of the output file name, its size and its date.

Example of output :
  2016/07/07 00:17:55 aloha alb3100 8.0.3 none save-etc: archive: |gzip: stdout: No space left on device | status: 1
  2016/07/07 00:17:55 aloha alb3100 8.0.3 none save-etc: cancel: rm empty config.new: status: 0
  2016/07/07 00:35:24 aloha alb3100 8.0.3 pts/0 save-etc: archive: |config.new|25464|2016-07-07+00:35:24 | status: 0

In order to limit the risk of accidents, the file is also marked for
append-only using "chattr +a".