scripts/restore-factory: update the flash's journal
Now we update the flash partition's journal when the config is restored
to factory defaults. The line indicates if the operation was performed
in interactive mode (implying the user responded Yes), and from what
terminal, or if it was called with "-f" in which case the parent process
name is also reported if possible, helping to know if it's from init or
any web interface. Example :

  2016/07/07 02:05:15 aloha alb3100 8.0.3 pts/0 restore-factory: confirmed in interactive mode
  2016/07/07 02:18:04 aloha alb3100 8.0.3 none restore-factory: called with -f from process cmdwrapper