scripts: improve config diff to support directories and removed files
There were several issues with the config diff. First, it used to
recurse into sub-directories regardless of the fact that -r was not
specified to diff. Second, removed files wouldn't be printed, which
is problematic when comparing with the flash as it can reveal an
operator mistake.

The comparison has now changed to use grep instead of tar to search
for files to be ignored. We now dump everything but directories in
the output to fix the double outputs and to ensure that even missing
files are listed.