scripts/restore-etc: do not mount a tmpfs on /etc if / is a ramfs
When booting an initramfs, it's pointless to mount some tmpfs everywhere
since we can already extract /etc into the / FS. We now detect that / is
a ramfs and avoid mounting the tmpfs in this case. At unmount time, we
remove the whole /etc entries if we're on a ramfs.

Note that we use stat to detect the filesystem type on /, and the script
ensures the tmpfs is still used if the "stat" binary is not present.