scripts/restore-etc: look for default settings in /usr/share/factory/settings
This file, when it exists, contains settings in the following format :

   ${program}/${setting}=${value}

Here restore-etc checks there for the file-system type (fstype) and the
mount options (mountopt). These ones can override the script's defaults,
but may be overridden by the command line.

This will particularly allow some models to override the default size
for /etc without having to patch the script nor call it with a command
line from a script where it's not convenient.