scripts/restore-etc: add support for a quiet mode with -q
Sometimes we don't want to see operations in progress. "config diff"
which calls "restore-etc -d" is a good example of this:

  # config diff
  Mounting /dev/mmcblk1p1 on /flash...
  Loading /flash/cfg/config.cur

And similarly just calling restore-etc with no option indicates whether
there are unsaved changes. Now when passing "-q" these messages will not
appear anymore. The "-d" option implies it.